requirements.txt
flask==2.0.1
flask-socketio==5.1.1
frida-tools==10.0.0
adb-shell>=0.4.0
requests==2.26.0
flask-cors==3.0.10
python-dotenv==0.19.1
pexpect==4.8.0
psutil==5.8.0
werkzeug==2.0.1 